requirements.txt
pyyaml>=6.0
Xlib>=0.21
screeninfo>=0.8
